Move metadata concurrency to a separate file
